Update dependencies, fix linting issues, split interfaces and default objects to resolve dependency cycles
