+++ package.json
@@ -13,7 +13,7 @@
     "cypress:open": "cypress open",
     "cypress:ci": "cypress run --record --group $GITHUB_REF --ci-build-id $GITHUB_SHA",
     "test": "run-s test:unit test:e2e",
-    "test:unit": "TS_NODE_TRANSPILE_ONLY=true NODE_ENV=test mocha",
+    "test:unit": "cross-env TS_NODE_TRANSPILE_ONLY=true NODE_ENV=test mocha",
     "test:unit:watch": "npm run test:unit -- --watch --inspect=5556",
     "test:unit:coverage": "NODE_ENV=test nyc --reporter=lcov --reporter=text --reporter=text-summary mocha",
     "test:e2e": "run-p --race start cypress:run",
@@ -53,56 +53,57 @@
     "js"
   ],
   "devDependencies": {
-    "@babel/core": "^7.6.4",
-    "@babel/preset-env": "^7.6.3",
-    "@babel/register": "^7.6.2",
-    "@types/chai": "^4.2.7",
-    "@types/mocha": "^5.2.7",
-    "@types/sinon": "^7.5.1",
-    "@types/sinon-chai": "^3.2.3",
-    "@typescript-eslint/eslint-plugin": "^2.11.0",
-    "@typescript-eslint/parser": "^2.11.0",
-    "autoprefixer": "^9.6.5",
-    "babel-loader": "^8.0.6",
-    "bundlesize": "^0.18.0",
-    "chai": "^4.2.0",
-    "cross-env": "^6.0.3",
+    "@babel/core": "^7.16.5",
+    "@babel/preset-env": "^7.16.5",
+    "@babel/register": "^7.16.5",
+    "@types/chai": "^4.3.0",
+    "@types/mocha": "^9.0.0",
+    "@types/sinon": "^10.0.6",
+    "@types/sinon-chai": "^3.2.6",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
+    "autoprefixer": "^10.4.0",
+    "babel-loader": "^8.2.3",
+    "bundlesize": "^0.18.1",
+    "chai": "^4.3.4",
+    "cross-env": "^7.0.3",
     "csso-cli": "^3.0.0",
-    "cypress": "3.6.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-config-prettier": "^6.5.0",
+    "cypress": "9.1.1",
+    "eslint": "^8.4.1",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^16.1.0",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-loader": "^3.0.2",
-    "eslint-plugin-compat": "3.3.0",
-    "eslint-plugin-cypress": "^2.8.1",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-sort-class-members": "^1.6.0",
-    "express": "^4.16.4",
-    "husky": "^3.0.9",
-    "jsdom": "^15.2.0",
-    "lint-staged": "^9.4.2",
-    "mocha": "^6.2.2",
-    "node-sass": "^4.12.0",
-    "nodemon": "^1.18.10",
+    "eslint-plugin-compat": "4.0.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-sort-class-members": "^1.14.1",
+    "express": "^4.17.1",
+    "husky": "^7.0.4",
+    "jsdom": "^19.0.0",
+    "lint-staged": "^12.1.2",
+    "mocha": "^9.1.3",
+    "node-sass": "^7.0.0",
+    "nodemon": "^2.0.15",
     "npm-run-all": "^4.1.5",
-    "nyc": "^14.1.1",
-    "postcss-cli": "^6.1.3",
-    "prettier": "^1.19.1",
-    "sinon": "^7.5.0",
-    "sinon-chai": "^3.3.0",
-    "ts-loader": "^6.2.1",
-    "ts-node": "^8.5.4",
-    "typescript": "^3.7.3",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-middleware": "^3.7.2",
-    "webpack-hot-middleware": "^2.25.0"
+    "nyc": "^15.1.0",
+    "postcss-cli": "^9.1.0",
+    "prettier": "^2.5.1",
+    "sinon": "^12.0.1",
+    "sinon-chai": "^3.7.0",
+    "ts-loader": "^9.2.6",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.4",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-middleware": "^5.2.2",
+    "webpack-hot-middleware": "^2.25.1"
   },
   "dependencies": {
-    "deepmerge": "^4.2.0",
+    "deepmerge": "^4.2.2",
     "fuse.js": "^3.4.6",
-    "redux": "^4.0.4"
+    "redux": "^4.1.2"
   },
   "npmName": "choices.js",
   "npmFileMap": [
@@ -125,7 +126,7 @@
   "bundlesize": [
     {
       "path": "public/assets/scripts/choices.min.js",
-      "maxSize": "20 kB"
+      "maxSize": "25 kB"
     },
     {
       "path": "public/assets/styles/choices.min.css",
