+++ public/assets/styles/base.css
@@ -78,8 +78,7 @@ a:focus {
   border-radius: 2.5px;
   font-size: 14px;
   -webkit-appearance: none;
-  -moz-appearance: none;
-       appearance: none;
+  appearance: none;
   margin-bottom: 24px;
 }
