+++ src/scripts/interfaces/action-type.ts
@@ -1,0 +1,12 @@
+export type ActionType =
+  | 'ADD_CHOICE'
+  | 'FILTER_CHOICES'
+  | 'ACTIVATE_CHOICES'
+  | 'CLEAR_CHOICES'
+  | 'ADD_GROUP'
+  | 'ADD_ITEM'
+  | 'REMOVE_ITEM'
+  | 'HIGHLIGHT_ITEM'
+  | 'CLEAR_ALL'
+  | 'RESET_TO'
+  | 'SET_IS_LOADING';
