+++ src/scripts/interfaces/event-type.ts
@@ -1,0 +1,11 @@
+export type EventType =
+  | 'addItem'
+  | 'removeItem'
+  | 'highlightItem'
+  | 'unhighlightItem'
+  | 'choice'
+  | 'change'
+  | 'search'
+  | 'showDropdown'
+  | 'hideDropdown'
+  | 'highlightChoice';
