+++ src/scripts/interfaces/group.ts
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+export interface Group {
+  id?: number;
+  active?: boolean;
+  disabled?: boolean;
+  value: any;
+}
