+++ src/scripts/interfaces/item.ts
@@ -1,0 +1,6 @@
+import { Choice } from './choice';
+
+export interface Item extends Choice {
+  choiceId?: number;
+  highlighted?: boolean;
+}
