+++ src/scripts/interfaces/keycode-map.ts
@@ -1,0 +1,11 @@
+export interface KeyCodeMap {
+  BACK_KEY: 46;
+  DELETE_KEY: 8;
+  ENTER_KEY: 13;
+  A_KEY: 65;
+  ESC_KEY: 27;
+  UP_KEY: 38;
+  DOWN_KEY: 40;
+  PAGE_UP_KEY: 33;
+  PAGE_DOWN_KEY: 34;
+}
