+++ src/scripts/interfaces/notice.ts
@@ -1,0 +1,5 @@
+// @todo rename
+export interface Notice {
+  response: boolean;
+  notice: string;
+}
