+++ src/scripts/interfaces/options.ts
@@ -1,261 +1,9 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { FuseOptions } from 'fuse.js';
-import Choices from './choices';
-
-export namespace Types {
-  export type strToEl = (
-    str: string,
-  ) => HTMLElement | HTMLInputElement | HTMLOptionElement;
-  export type stringFunction = () => string;
-  export type noticeStringFunction = (value: string) => string;
-  export type noticeLimitFunction = (maxItemCount: number) => string;
-  export type filterFunction = (value: string) => boolean;
-  export type valueCompareFunction = (
-    value1: string,
-    value2: string,
-  ) => boolean;
-}
-
-export interface Choice {
-  id?: number;
-  customProperties?: Record<string, any>;
-  disabled?: boolean;
-  active?: boolean;
-  elementId?: number;
-  groupId?: number;
-  keyCode?: number;
-  label: string;
-  placeholder?: boolean;
-  selected?: boolean;
-  value: string;
-  score?: number;
-  choices?: Choice[];
-}
-
-export interface Group {
-  id?: number;
-  active?: boolean;
-  disabled?: boolean;
-  value: any;
-}
-export interface Item extends Choice {
-  choiceId?: number;
-  highlighted?: boolean;
-}
-
-/**
- * Events fired by Choices behave the same as standard events. Each event is triggered on the element passed to Choices (accessible via `this.passedElement`. Arguments are accessible within the `event.detail` object.
- */
-export interface EventMap {
-  /**
-   * Triggered each time an item is added (programmatically or by the user).
-   *
-   * **Input types affected:** text, select-one, select-multiple
-   *
-   * Arguments: id, value, label, groupValue, keyCode
-   */
-  addItem: CustomEvent<{
-    id: number;
-    value: string;
-    label: string;
-    groupValue: string;
-    keyCode: number;
-  }>;
-
-  /**
-   * Triggered each time an item is removed (programmatically or by the user).
-   *
-   * **Input types affected:** text, select-one, select-multiple
-   *
-   * Arguments: id, value, label, groupValue
-   */
-  removeItem: CustomEvent<{
-    id: number;
-    value: string;
-    label: string;
-    groupValue: string;
-  }>;
-
-  /**
-   * Triggered each time an item is highlighted.
-   *
-   * **Input types affected:** text, select-multiple
-   *
-   * Arguments: id, value, label, groupValue
-   */
-  highlightItem: CustomEvent<{
-    id: number;
-    value: string;
-    label: string;
-    groupValue: string;
-  }>;
-
-  /**
-   * Triggered each time an item is unhighlighted.
-   *
-   * **Input types affected:** text, select-multiple
-   *
-   * Arguments: id, value, label, groupValue
-   */
-  unhighlightItem: CustomEvent<{
-    id: number;
-    value: string;
-    label: string;
-    groupValue: string;
-  }>;
-
-  /**
-   * Triggered each time a choice is selected **by a user**, regardless if it changes the value of the input.
-   *
-   * **Input types affected:** select-one, select-multiple
-   *
-   * Arguments: choice: Choice
-   */
-  choice: CustomEvent<{ choice: Choice }>;
-
-  /**
-   * Triggered each time an item is added/removed **by a user**.
-   *
-   * **Input types affected:** text, select-one, select-multiple
-   *
-   * Arguments: value
-   */
-  change: CustomEvent<{ value: string }>;
-
-  /**
-   * Triggered when a user types into an input to search choices.
-   *
-   * **Input types affected:** select-one, select-multiple
-   *
-   * Arguments: value, resultCount
-   */
-  search: CustomEvent<{ value: string; resultCount: number }>;
-
-  /**
-   * Triggered when the dropdown is shown.
-   *
-   * **Input types affected:** select-one, select-multiple
-   *
-   * Arguments: -
-   */
-  showDropdown: CustomEvent<undefined>;
-
-  /**
-   * Triggered when the dropdown is hidden.
-   *
-   * **Input types affected:** select-one, select-multiple
-   *
-   * Arguments: -
-   */
-  hideDropdown: CustomEvent<undefined>;
-
-  /**
-   * Triggered when a choice from the dropdown is highlighted.
-   *
-   * Input types affected: select-one, select-multiple
-   * Arguments: el is the choice.passedElement that was affected.
-   */
-  highlightChoice: CustomEvent<{ el: PassedElement }>;
-}
-
-export interface KeyCodeMap {
-  BACK_KEY: 46;
-  DELETE_KEY: 8;
-  ENTER_KEY: 13;
-  A_KEY: 65;
-  ESC_KEY: 27;
-  UP_KEY: 38;
-  DOWN_KEY: 40;
-  PAGE_UP_KEY: 33;
-  PAGE_DOWN_KEY: 34;
-}
-
-export type ActionType =
-  | 'ADD_CHOICE'
-  | 'FILTER_CHOICES'
-  | 'ACTIVATE_CHOICES'
-  | 'CLEAR_CHOICES'
-  | 'ADD_GROUP'
-  | 'ADD_ITEM'
-  | 'REMOVE_ITEM'
-  | 'HIGHLIGHT_ITEM'
-  | 'CLEAR_ALL'
-  | 'RESET_TO'
-  | 'SET_IS_LOADING';
-
-/** Classes added to HTML generated by  By default classnames follow the BEM notation. */
-export interface ClassNames {
-  /** @default 'choices' */
-  containerOuter: string;
-  /** @default 'choices__inner' */
-  containerInner: string;
-  /** @default 'choices__input' */
-  input: string;
-  /** @default 'choices__input--cloned' */
-  inputCloned: string;
-  /** @default 'choices__list' */
-  list: string;
-  /** @default 'choices__list--multiple' */
-  listItems: string;
-  /** @default 'choices__list--single' */
-  listSingle: string;
-  /** @default 'choices__list--dropdown' */
-  listDropdown: string;
-  /** @default 'choices__item' */
-  item: string;
-  /** @default 'choices__item--selectable' */
-  itemSelectable: string;
-  /** @default 'choices__item--disabled' */
-  itemDisabled: string;
-  /** @default 'choices__item--choice' */
-  itemChoice: string;
-  /** @default 'choices__placeholder' */
-  placeholder: string;
-  /** @default 'choices__group' */
-  group: string;
-  /** @default 'choices__heading' */
-  groupHeading: string;
-  /** @default 'choices__button' */
-  button: string;
-  /** @default 'is-active' */
-  activeState: string;
-  /** @default 'is-focused' */
-  focusState: string;
-  /** @default 'is-open' */
-  openState: string;
-  /** @default 'is-disabled' */
-  disabledState: string;
-  /** @default 'is-highlighted' */
-  highlightedState: string;
-  /** @default 'is-selected' */
-  selectedState: string;
-  /** @default 'is-flipped' */
-  flippedState: string;
-  /** @default 'is-loading' */
-  loadingState: string;
-  /** @default 'has-no-results' */
-  noResults: string;
-  /** @default 'has-no-choices' */
-  noChoices: string;
-}
-
-export interface PassedElement extends HTMLElement {
-  classNames: ClassNames;
-  element: (HTMLInputElement | HTMLSelectElement) & {
-    // Extends HTMLElement addEventListener with Choices events
-    addEventListener<K extends keyof EventMap>(
-      type: K,
-      listener: (
-        this: HTMLInputElement | HTMLSelectElement,
-        ev: EventMap[K],
-      ) => void,
-      options?: boolean | AddEventListenerOptions,
-    ): void;
-  };
-  type: 'text' | 'select-one' | 'select-multiple';
-  isDisabled: boolean;
-  parentInstance: Choices;
-}
+import { Choices } from './choices';
+import { Choice } from './choice';
+import { ClassNames } from './class-names';
+import { PositionOptionsType } from './position-options-type';
+import { Types } from './types';
 
 /**
  * Choices options interface
@@ -370,7 +118,7 @@ export interface Options {
    *
    * @default null
    */
-  addItemFilter: string | RegExp | Types.filterFunction | null;
+  addItemFilter: string | RegExp | Types.FilterFunction | null;
 
   /**
    * The text that is shown when a user has inputted a new item but has not pressed the enter key. To access the current input value, pass a function with a `value` argument (see the **default config** [https://github.com/jshjohnson/Choices#setup] for an example), otherwise pass a string.
@@ -382,7 +130,7 @@ export interface Options {
    * (value) => `Press Enter to add <b>"${value}"</b>`;
    * ```
    */
-  addItemText: string | Types.noticeStringFunction;
+  addItemText: string | Types.NoticeStringFunction;
 
   /**
    * Whether a user can remove items.
@@ -492,7 +240,7 @@ export interface Options {
    *
    * @default 'auto'
    */
-  position: 'auto' | 'top' | 'bottom';
+  position: PositionOptionsType;
 
   /**
    * Whether the scroll position should reset after adding an item.
@@ -620,7 +368,7 @@ export interface Options {
    *
    * @default 'No results found'
    */
-  noResultsText: string | Types.stringFunction;
+  noResultsText: string | Types.StringFunction;
 
   /**
    * The text that is shown when a user has selected all possible choices. Optionally pass a function returning a string.
@@ -629,7 +377,7 @@ export interface Options {
    *
    * @default 'No choices to choose from'
    */
-  noChoicesText: string | Types.stringFunction;
+  noChoicesText: string | Types.StringFunction;
 
   /**
    * The text that is shown when a user hovers over a selectable choice.
@@ -650,14 +398,14 @@ export interface Options {
    * (maxItemCount) => `Only ${maxItemCount} values can be added.`;
    * ```
    */
-  maxItemText: string | Types.noticeLimitFunction;
+  maxItemText: string | Types.NoticeLimitFunction;
 
   /**
    * If no duplicates are allowed, and the value already exists in the array.
    *
    * @default 'Only unique values can be added'
    */
-  uniqueItemText: string | Types.noticeStringFunction;
+  uniqueItemText: string | Types.NoticeStringFunction;
 
   /**
    * The text that is shown when addItemFilter is passed and it returns false
@@ -666,7 +414,7 @@ export interface Options {
    *
    * @default 'Only values matching specific conditions can be added'
    */
-  customAddItemText: string | Types.noticeStringFunction;
+  customAddItemText: string | Types.NoticeStringFunction;
 
   /**
    * Compare choice and value in appropriate way (e.g. deep equality for objects). To compare choice and value, pass a function with a `valueComparer` argument (see the [default config](https://github.com/jshjohnson/Choices#setup) for an example).
@@ -678,7 +426,7 @@ export interface Options {
    * (choice, item) => choice === item;
    * ```
    */
-  valueComparer: Types.valueCompareFunction;
+  valueComparer: Types.ValueCompareFunction;
 
   /**
    * Classes added to HTML generated by  By default classnames follow the BEM notation.
@@ -737,18 +485,5 @@ export interface Options {
    *
    * @default null
    */
-  callbackOnCreateTemplates: ((template: Types.strToEl) => void) | null;
-}
-
-// @todo rename
-export interface Notice {
-  response: boolean;
-  notice: string;
-}
-
-export interface State {
-  choices: Choice[];
-  groups: Group[];
-  items: Item[];
-  loading: boolean;
+  callbackOnCreateTemplates: ((template: Types.StrToEl) => void) | null;
 }
