+++ src/scripts/interfaces/passed-element-type.ts
@@ -1,0 +1,1 @@
+export type PassedElementType = 'text' | 'select-one' | 'select-multiple';
