+++ src/scripts/interfaces/position-options-type.ts
@@ -1,0 +1,1 @@
+export type PositionOptionsType = 'auto' | 'top' | 'bottom';
