+++ src/scripts/store/store.ts
@@ -1,7 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { createStore, Store as IStore, AnyAction } from 'redux';
+import { Choice } from '../interfaces/choice';
+import { Group } from '../interfaces/group';
+import { Item } from '../interfaces/item';
+import { State } from '../interfaces/state';
 import rootReducer from '../reducers/index';
-import { Choice, Group, Item, State } from '../interfaces';
 
 export default class Store {
   _store: IStore;
@@ -46,14 +49,14 @@ export default class Store {
    * Get active items from store
    */
   get activeItems(): Item[] {
-    return this.items.filter(item => item.active === true);
+    return this.items.filter((item) => item.active === true);
   }
 
   /**
    * Get highlighted items from store
    */
   get highlightedActiveItems(): Item[] {
-    return this.items.filter(item => item.active && item.highlighted);
+    return this.items.filter((item) => item.active && item.highlighted);
   }
 
   /**
@@ -67,21 +70,23 @@ export default class Store {
    * Get active choices from store
    */
   get activeChoices(): Choice[] {
-    return this.choices.filter(choice => choice.active === true);
+    return this.choices.filter((choice) => choice.active === true);
   }
 
   /**
    * Get selectable choices from store
    */
   get selectableChoices(): Choice[] {
-    return this.choices.filter(choice => choice.disabled !== true);
+    return this.choices.filter((choice) => choice.disabled !== true);
   }
 
   /**
    * Get choices that can be searched (excluding placeholders)
    */
   get searchableChoices(): Choice[] {
-    return this.selectableChoices.filter(choice => choice.placeholder !== true);
+    return this.selectableChoices.filter(
+      (choice) => choice.placeholder !== true,
+    );
   }
 
   /**
@@ -90,7 +95,7 @@ export default class Store {
   get placeholderChoice(): Choice | undefined {
     return [...this.choices]
       .reverse()
-      .find(choice => choice.placeholder === true);
+      .find((choice) => choice.placeholder === true);
   }
 
   /**
@@ -106,10 +111,10 @@ export default class Store {
   get activeGroups(): Group[] {
     const { groups, choices } = this;
 
-    return groups.filter(group => {
+    return groups.filter((group) => {
       const isActive = group.active === true && group.disabled === false;
       const hasActiveOptions = choices.some(
-        choice => choice.active === true && choice.disabled === false,
+        (choice) => choice.active === true && choice.disabled === false,
       );
 
       return isActive && hasActiveOptions;
@@ -127,13 +132,13 @@ export default class Store {
    * Get single choice by it's ID
    */
   getChoiceById(id: string): Choice | undefined {
-    return this.activeChoices.find(choice => choice.id === parseInt(id, 10));
+    return this.activeChoices.find((choice) => choice.id === parseInt(id, 10));
   }
 
   /**
    * Get group by group id
    */
   getGroupById(id: number): Group | undefined {
-    return this.groups.find(group => group.id === id);
+    return this.groups.find((group) => group.id === id);
   }
 }
